+++ tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(calcofi4r)
+
+test_check("calcofi4r")